+++ Assets/Scripts/ValtraRecord.cs
@@ -3,7 +3,7 @@ using System;
 [Serializable]
 public class ValtraRecord
 {
-    public float gpsTime;
+    public double gpsTime;
     public double easting;
     public double northing;
 }
